Moved summary cards to separate folder
