+++ upnext-web-app/src/UpNext/components/SummaryCards/BookSummaryCard.tsx
