+++ upnext-web-app/src/UpNext/components/SummaryCards/GameSummaryCard.tsx
